\name{AMKATpackage}
\alias{AMKATpackage}
\docType{package}

\title{
\packageTitle{AMKAT}
}

\description{
\packageDescription{AMKAT}
}

\details{
The main function \code{amkat} performs the test. All other functions are either helper functions used by \code{amkat} that have been made accessible to increase the package's utility and flexibility (e.g., \code{generateKernelMatrix} and \code{applyAmkatFilter}), or functions provided to assist in the evaluation of AMKAT's methodology. A complete index of help topics for the functions contained in the package is included below.

\packageIndices{AMKAT}

The DESCRIPTION file:
\packageDESCRIPTION{AMKAT}
}

\author{
\packageAuthor{AMKAT}

Maintainer: \packageMaintainer{AMKAT}
}

\references{Neal, Brian and He, Tao. \dQuote{An adaptive multivariate kernel-based test for association with multiple quantitative traits in high-dimensional data.} \emph{Genetic Epidemiology} (not yet submitted).}

\keyword{ package }
